changeover to local (relative) coordinate system also: link controls moved & drawn accurately at scale, link label starts at right location other misc cleanup
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@4336 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/tufts/vue/LWIcon.java
@@ -306,7 +306,7 @@ public abstract class LWIcon extends Rectangle2D.Float
                 Rectangle2D.Float tipRegion = (Rectangle2D.Float) tipIcon.getBounds2D();
                 if (mCoordsNodeLocal) {
                     // translate tipRegion from component to map coords
-                    float s = mLWC.getScale();
+                    double s = mLWC.getScale();
                     if (s != 1) {
                         tipRegion.x *= s;
                         tipRegion.y *= s;
